Update README.md
In this lab act using the if-elif-else I've created a program/code about classifying/categorizing in which group you belong. It is based on your age but if input a wrong number or it has negative it will give you a appropriate message. This program needs the age of the user to make it work.